Moved some classes to their respective packages
