+++ api/src/test/java/org/openmrs/module/patientgrid/filter/LocationCohortDefinitionEvaluatorTest.java
@@ -12,6 +12,7 @@ import org.openmrs.Location;
 import org.openmrs.Patient;
 import org.openmrs.api.LocationService;
 import org.openmrs.module.datafilter.impl.api.DataFilterService;
+import org.openmrs.module.patientgrid.filter.definition.LocationCohortDefinition;
 import org.openmrs.module.reporting.cohort.EvaluatedCohort;
 import org.openmrs.module.reporting.cohort.definition.service.CohortDefinitionService;
 import org.openmrs.test.BaseModuleContextSensitiveTest;
