+++ api/src/test/java/org/openmrs/module/patientgrid/filter/ObsForLatestEncounterCohortDefinitionEvaluatorTest.java
@@ -9,6 +9,7 @@ import org.openmrs.Concept;
 import org.openmrs.EncounterType;
 import org.openmrs.Patient;
 import org.openmrs.api.context.Context;
+import org.openmrs.module.patientgrid.filter.definition.ObsForLatestEncounterCohortDefinition;
 import org.openmrs.module.reporting.cohort.EvaluatedCohort;
 import org.openmrs.module.reporting.cohort.definition.service.CohortDefinitionService;
 import org.openmrs.test.BaseModuleContextSensitiveTest;
