+++ api/src/test/java/org/openmrs/module/patientgrid/filter/PatientGridFilterUtilsTest.java
@@ -38,6 +38,9 @@ import org.openmrs.module.patientgrid.PatientGrid;
 import org.openmrs.module.patientgrid.PatientGridColumn;
 import org.openmrs.module.patientgrid.PatientGridColumn.ColumnDatatype;
 import org.openmrs.module.patientgrid.PatientGridColumnFilter;
+import org.openmrs.module.patientgrid.filter.definition.AgeRangeAtLatestEncounterCohortDefinition;
+import org.openmrs.module.patientgrid.filter.definition.LocationCohortDefinition;
+import org.openmrs.module.patientgrid.filter.definition.ObsForLatestEncounterCohortDefinition;
 import org.openmrs.module.reporting.cohort.definition.CompositionCohortDefinition;
 import org.openmrs.module.reporting.cohort.definition.GenderCohortDefinition;
 import org.openmrs.module.reporting.common.AgeRange;
